requirements.txt
fastapi==0.119.0
httpx==0.28.1
langchain==0.3.27
langchain_chroma==1.0.0
langchain_community==0.4
langchain_huggingface==1.0.0
openai==2.5.0
pandas==2.3.3
pydantic==2.12.3
pydantic_settings==2.11.0
streamlit==1.49.1
google-genai